% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{batter_projections}
\alias{batter_projections}
\title{Example batter and pitcher projections}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 592 rows and 51 columns.
}
\usage{
batter_projections
}
\description{
ATC ROS projections were scraped using gardy::fg_batter_projections() and
gardy::fg_pitcher_projections(). They are good representations of the raw
data that will be imported for use in the package. Included here for package
development and testing so the new projections do not need to be scraped each
session.
}
\keyword{datasets}
