% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pitcher_projections}
\alias{pitcher_projections}
\title{Example pitcher projections}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 809 rows and 41 columns.
}
\usage{
pitcher_projections
}
\description{
ATC ROS projections scraped from fangraphs.com in roughly mid-2024.
It is a good representation of the raw pitcher projection data on which the
package will be used. Included here for development and testing so new
projections do not need to be imported each session.
}
\keyword{datasets}
