% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpp.R
\name{lpp}
\alias{lpp}
\title{Run Last Player Picked}
\usage{
lpp(
  bat = NULL,
  pit = NULL,
  keepers = NULL,
  lg = "MLB",
  teams = 12,
  budget = 260,
  min_bid = 1,
  bat_cat = c("HR", "R", "RBI", "SB", "OBP"),
  pit_cat = c("WQS", "SVHLD", "SO", "ERA", "WHIP"),
  bat_pos = c(C = 1, `1B` = 1, `2B` = 1, `3B` = 1, SS = 1, CI = 1, MI = 1, OF = 5, UT =
    1),
  pit_pos = c(SP = 6, RP = 3, P = 0),
  bench = 2
)
}
\arguments{
\item{bat}{A data frame of batter projections.}

\item{pit}{A data frame of pitcher projections.}

\item{keepers}{A data frame of keepers and values. Defaults to NULL.}

\item{lg}{League: "AL", "NL", or "MLB" (the default).}

\item{teams}{An integer. The number of teams in the league. Defaults to 12.}

\item{budget}{An integer. The starting auction budget of each team in the league. Defaults to 260.}

\item{min_bid}{An integer. The minimum auction bid. Defaults to 1.}

\item{bat_cat}{A character vector of league batting categories.}

\item{pit_cat}{A character vector of league pitching categories.}

\item{bat_pos}{A named integer vector. The number of batters drafted at each position per team.}

\item{pit_pos}{A named integer vector. The number of pitchers drafted at each position per team.}

\item{bench}{An integer. The number of bench players drafted by each team.}
}
\value{
A list of length 2 containing a data frame of batter and pitcher auction values.
}
\description{
Run Last Player Picked
}
